v2-rag-from-scratch/pyproject.toml
[project]
name = "studybuddy"
version = "2.0.0"
description = "AI-powered study assistant - Chapter 2: RAG from scratch"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "openai>=1.54.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.26.0",
]
